deleteOldImage fonksiyonu değiştirildi. Rpute den kaynaklanan hata çözümü yapıldı ve silme işlemi başarılı bir şekilde gerçekleştirildi

+++ app/Support/helper.php
@@ -24,6 +24,11 @@ class Helper
     }
     public static function deleteOldImage(string $path)
     {
-        return unlink(public_path($path));
+        if (is_file($path))
+        {
+            return unlink(public_path($path));
+        }
+        else
+            return false;
     }
 }
